+++ src/cli.rs
@@ -1,73 +1,62 @@
-use crate::tre::{self, Coloring, Mode, RunOption};
-use regex::Regex;
+use clap::{ArgEnum, Parser};
 
-pub fn get_run_option(args: &[String]) -> RunOption {
-    match tre::cli_options().parse(&args[1..]) {
-        Ok(matches) => {
-            let mode: Mode = if matches.opt_present("v") {
-                Mode::Version
-            } else if matches.opt_present("h") {
-                Mode::Help
-            } else if matches.opt_present("a") {
-                Mode::ShowAllFiles
-            } else if !cfg!(windows) && matches.opt_present("s") {
-                Mode::ExcludeHiddenFiles
-            } else {
-                Mode::FollowGitIgnore
-            };
+#[derive(ArgEnum, Clone, Debug)]
+pub enum Coloring {
+    Automatic,
+    Always,
+    Never,
+}
 
-            let directories_only = matches.opt_present("d");
-            let output_json = matches.opt_present("j");
-            let editor: Option<Option<String>> = if matches.opt_present("e") {
-                Some(matches.opt_str("e"))
-            } else {
-                None
-            };
-            let root: Option<String> = if matches.free.is_empty() {
-                None
-            } else {
-                Some(matches.free[0].clone())
-            };
-
-            let max_depth: Option<usize> = matches.opt_get("l").unwrap_or(None);
-
-            let exclude_patterns = matches
-                .opt_strs("E")
-                .iter()
-                .map(|s| Regex::new(s).expect(""))
-                .collect();
-
-            let color_string: Option<String> = matches
-                .opt_get("c")
-                .unwrap_or(None)
-                .map(|s: String| s.to_lowercase());
-
-            let coloring: Coloring = match color_string.as_deref() {
-                Some("always") => Coloring::Always,
-                Some("never") => Coloring::Never,
-                _ => Coloring::Automatic,
-            };
-
-            RunOption {
-                mode,
-                editor,
-                directories_only,
-                output_json,
-                root,
-                max_depth,
-                exclude_patterns,
-                coloring,
-            }
-        }
-        Err(_) => RunOption {
-            mode: Mode::Help,
-            editor: None,
-            directories_only: false,
-            output_json: false,
-            root: None,
-            max_depth: None,
-            exclude_patterns: [].to_vec(),
-            coloring: Coloring::Automatic,
-        },
+impl core::str::FromStr for Coloring {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s.to_lowercase().as_str() {
+            "never" => Coloring::Never,
+            "always" => Coloring::Always,
+            _ => Coloring::Automatic,
+        })
     }
+}
+
+#[derive(Debug, Clone, Parser)]
+#[clap(author, version, about)]
+pub struct Interface {
+    /// Print all files and directories, including hidden ones.
+    #[clap(long, short, parse(from_flag))]
+    pub all: bool,
+    /// Use normal print despite gitignore settings. '-a' has higher priority.
+    #[cfg(not(target_os = "windows"))]
+    #[clap(long, short, parse(from_flag))]
+    pub simple: bool,
+    /// Only list directories in output.
+    #[clap(long, short, parse(from_flag))]
+    pub directories: bool,
+    /// Create aliases for each displayed result, and add a number in front of file name to
+    /// indicate the alias name. For example, a number "42" means an shell alias "e42" has been
+    /// created. Running "e42" will cause the associated file or directory to be open with $EDITOR
+    /// (or a default program for the file type on Windows), or a command specified along with this
+    /// command.
+    #[clap(long, short, value_name = "COMMAND")]
+    pub editor: Option<Option<String>>,
+    /// Output JSON instead of tree diagram.
+    #[clap(long, short, parse(from_flag))]
+    pub json: bool,
+    /// Limit depth of the tree in output.
+    #[clap(long, short)]
+    pub limit: Option<usize>,
+    /// Exclude paths matching a regex pattern. Repeatable.
+    #[clap(long, short = 'E', value_name = "PATTERN")]
+    pub exclude: Vec<String>,
+    #[clap(
+        arg_enum,
+        long,
+        short,
+        value_name = "WHEN",
+        default_value = "automatic"
+    )]
+    /// When to color the output. `automatic` means when printing to a terminal, tre will include
+    /// colors; otherwise it will disable colors.
+    pub color: Coloring,
+    #[clap(default_value = ".")]
+    pub path: String,
 }