+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support installing from Nix flake.
 - Add `-E/--exclude PATTERN` option to exclude paths from results.
 - Add `-e/--color WHEN` option to control colors in output.
+- The short flag for `--version` is renamed from `-v` to `-V`.
 
 # 0.3.6